# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VCCDB HTTP API
  version: 0.1.0
  description: |
    HTTP API für VCC Multi‑Modell‑Datenbank. Beinhaltet Health, CRUD, Indexverwaltung, Query,
    und Observability (/stats, /metrics). Die Query-Engine unterstützt AND-Gleichheitsprädikate,
    Range-Prädikate (gte/lte) und ORDER BY über Range-Indizes. Optional: Fallback Full-Scan.
servers:
  - url: http://localhost:8765
paths:
...
  /content/import:
    post:
      summary: Content importieren (v0 Bulk)
      operationId: contentImport
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentImportRequest'
      responses:
        '200':
          description: Import erfolgreich
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentImportResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  schemas:
    ContentMeta:
      type: object
      properties:
        id: { type: string }
        mime_type: { type: string }
        category: { type: integer, description: "Enum: TEXT=0, IMAGE=1, AUDIO=2, VIDEO=3, GEO=4, CAD=5, ARCHIVE=6, STRUCTURED=7, BINARY=8, UNKNOWN=9" }
        original_filename: { type: string }
        size_bytes: { type: integer, format: int64 }
        created_at: { type: integer, format: int64 }
        modified_at: { type: integer, format: int64 }
        hash_sha256: { type: string }
        text_extracted: { type: boolean }
        chunked: { type: boolean }
        indexed: { type: boolean }
        chunk_count: { type: integer }
        embedding_dim: { type: integer }
        extracted_metadata: { type: object, additionalProperties: true }
        user_metadata: { type: object, additionalProperties: true }
        tags:
          type: array
          items: { type: string }
        parent_id: { type: string }
        child_ids:
          type: array
          items: { type: string }
    ChunkMeta:
      type: object
      properties:
        id: { type: string }
        content_id: { type: string }
        seq_num: { type: integer }
        chunk_type: { type: string }
        text: { type: string }
        data: { type: object, additionalProperties: true }
        blob_ref: { type: string }
        start_offset: { type: integer }
        end_offset: { type: integer }
        embedding:
          type: array
          items: { type: number }
        embedding_indexed: { type: boolean }
        created_at: { type: integer, format: int64 }
    ContentImportRequest:
      type: object
      properties:
        content: { $ref: '#/components/schemas/ContentMeta' }
        chunks:
          type: array
          items: { $ref: '#/components/schemas/ChunkMeta' }
        edges:
          type: array
          items: { type: object, additionalProperties: true }
        blob: { type: string, description: "Roh-Blob als String" }
        blob_base64: { type: string, description: "Base64-kodierter Blob (im MVP nicht automatisch dekodiert)" }
      required: [content]
    ContentImportResponse:
      type: object
      properties:
        status: { type: string }
        content_id: { type: string }
      required: [status, content_id]
...
  responses:
    BadRequest:
      description: Ungültige Anforderung
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Interner Serverfehler
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: boolean }
        message: { type: string }
        status_code: { type: integer }
      required: [error, message, status_code]
...